requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
diskcache>=5.6.3
python-dotenv>=1.0.0
lxml>=4.9.0
parsel>=1.8.1
chardet>=5.2.0
brotli>=1.1.0
loguru>=0.7.0
undetected-chromedriver>=3.5.4
selenium>=4.15.0